requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
binance-connector==1.13.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
dateparser==1.1.1
decorator==5.1.1
executing==0.8.3
Flask==2.1.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
ipdb==0.13.9
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multidict==6.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
python-binance==1.0.16
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.0
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.39
stack-data==0.3.0
toml==0.10.2
traitlets==5.3.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ujson==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
zope.interface==5.4.0
